+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [1.8.1] - 2018-09-14
 
+### Changed
+
+- We've enabled the Preflight experiment by default
+
 ### Fixed
 
 - Fixed incorrect version in the extension settings tab